+++ contracts/Org.sol
@@ -28,10 +28,10 @@ contract Org is Administratable {
         bool filesSubmitted;
     }
 
-    uint public taxId;
+    uint256 public taxId;
     address public orgWallet;
     Claim[] public claims;
-    event cashOutComplete(uint cashOutAmount);
+    event cashOutComplete(uint256 cashOutAmount);
 
 
 // ========== CONSTRUCTOR ==========    
@@ -41,7 +41,7 @@ contract Org is Administratable {
     * @param ein The U.S. Tax Identification Number for the Organization
     * @param adminContractAddress Contract Address for Endaoment Admin
     */
-    constructor(uint ein, address adminContractAddress) public onlyAdminOrRole(adminContractAddress, IEndaomentAdmin.Role.ORG_FACTORY){
+    constructor(uint256 ein, address adminContractAddress) public onlyAdminOrRole(adminContractAddress, IEndaomentAdmin.Role.ORG_FACTORY){
         taxId = ein;
     }
 
@@ -75,7 +75,7 @@ contract Org is Administratable {
      * @param  index Index value of Claim.
      * @param adminContractAddress Contract Address for Endaoment Admin
      */
-    function approveClaim(uint index, address adminContractAddress) public onlyAdminOrRole(adminContractAddress, IEndaomentAdmin.Role.REVIEWER){
+    function approveClaim(uint256 index, address adminContractAddress) public onlyAdminOrRole(adminContractAddress, IEndaomentAdmin.Role.REVIEWER){
         Claim storage claim = claims[index]; 
         
         setOrgWallet(claim.desiredWallet, adminContractAddress);
